Split CityGML bootstrap model files
